refactor: [step3] FinalFrames.of(10)을 FinalFrames.of(Frames.LAST_FRAME) 으로 변경
